requirements.txt
django
djangorestframework==3.3.1
